# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * HireLens AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type InterviewInputDifficulty = typeof InterviewInputDifficulty[keyof typeof InterviewInputDifficulty];


export const InterviewInputDifficulty = {
  Easy: 'Easy',
  Medium: 'Medium',
  Hard: 'Hard',
} as const;

export type InterviewInputType = typeof InterviewInputType[keyof typeof InterviewInputType];


export const InterviewInputType = {
  Technical: 'Technical',
  HR: 'HR',
} as const;

export interface InterviewInput {
  role: string;
  difficulty: InterviewInputDifficulty;
  type: InterviewInputType;
}

export interface InterviewQuestion {
  id: string;
  prompt: string;
  category: string;
  answered: boolean;
}

export type InterviewStatus = typeof InterviewStatus[keyof typeof InterviewStatus];


export const InterviewStatus = {
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export interface Interview {
  id: string;
  role: string;
  difficulty: string;
  type: string;
  status: InterviewStatus;
  /** @nullable */
  score: number | null;
  questionCount: number;
  /** @nullable */
  completedAt: string | null;
  createdAt: string;
  questions: InterviewQuestion[];
}

export interface AnswerInput {
  questionId: string;
  answer: string;
}

export interface AnswerFeedback {
  score: number;
  good: string[];
  improve: string[];
  betterAnswer: string;
  /** @nullable */
  nextQuestion: string | null;
  completed: boolean;
}

export interface WeakArea {
  subject: string;
  score: number;
  fullMark: number;
}

export interface DashboardSummary {
  interviewsTaken: number;
  averageScore: number;
  practiceMinutes: number;
  improvement: number;
  weakAreas: WeakArea[];
  recentInterviews: Interview[];
}

export interface ResumeAnalysisInput {
  filename: string;
  text: string;
}

export interface ResumeAnalysis {
  id: string;
  filename: string;
  atsScore: number;
  summary: string;
  missingSkills: string[];
  strengths: string[];
  questions: string[];
}

export interface AdminUser {
  id: string;
  name: string;
  email: string;
  plan: string;
  interviews: number;
  joinedAt: string;
  status: string;
}