# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * HireLens AI API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUser {
  id: string;
  name: string;
  email: string;
  plan: string;
  interviews: number;
  joinedAt: string;
  status: string;
}