# lib/api-zod/src/generated/types/answerFeedback.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * HireLens AI API
 * OpenAPI spec version: 0.1.0
 */

export interface AnswerFeedback {
  score: number;
  good: string[];
  improve: string[];
  betterAnswer: string;
  /** @nullable */
  nextQuestion: string | null;
  completed: boolean;
}