# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * HireLens AI API
 * OpenAPI spec version: 0.1.0
 */
import type { Interview } from './interview';
import type { WeakArea } from './weakArea';

export interface DashboardSummary {
  interviewsTaken: number;
  averageScore: number;
  practiceMinutes: number;
  improvement: number;
  weakAreas: WeakArea[];
  recentInterviews: Interview[];
}