# lib/api-zod/src/generated/types/interview.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * HireLens AI API
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewQuestion } from './interviewQuestion';
import type { InterviewStatus } from './interviewStatus';

export interface Interview {
  id: string;
  role: string;
  difficulty: string;
  type: string;
  status: InterviewStatus;
  /** @nullable */
  score: number | null;
  questionCount: number;
  /** @nullable */
  completedAt: string | null;
  createdAt: string;
  questions: InterviewQuestion[];
}