# lib/api-zod/src/generated/types/interviewInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * HireLens AI API
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewInputDifficulty } from './interviewInputDifficulty';
import type { InterviewInputType } from './interviewInputType';

export interface InterviewInput {
  role: string;
  difficulty: InterviewInputDifficulty;
  type: InterviewInputType;
}