# lib/api-zod/src/generated/types/interviewQuestion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * HireLens AI API
 * OpenAPI spec version: 0.1.0
 */

export interface InterviewQuestion {
  id: string;
  prompt: string;
  category: string;
  answered: boolean;
}