# lib/api-zod/src/generated/types/resumeAnalysis.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * HireLens AI API
 * OpenAPI spec version: 0.1.0
 */

export interface ResumeAnalysis {
  id: string;
  filename: string;
  atsScore: number;
  summary: string;
  missingSkills: string[];
  strengths: string[];
  questions: string[];
}