# lib/api-zod/src/generated/types/weakArea.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * HireLens AI API
 * OpenAPI spec version: 0.1.0
 */

export interface WeakArea {
  subject: string;
  score: number;
  fullMark: number;
}